+++ src/index.ts
@@ -226,6 +226,7 @@ export { mustBeObject } from "./BasicTypes/Objects/mustBeObject";
 export { mustBeObjectish } from "./BasicTypes/Objects/mustBeObjectish";
 export type { NonNullableObject } from "./BasicTypes/Objects/NonNullObject.type";
 export type { ObjectKey } from "./BasicTypes/Objects/ObjectKey.type";
+export { omitProperties } from "./BasicTypes/Objects/omitProperties";
 export { pickProperties } from "./BasicTypes/Objects/pickProperties";
 export { setProperty } from "./BasicTypes/Objects/setProperty";
 export { validateObject } from "./BasicTypes/Objects/validateObject";